+++ lib/examples.rb
@@ -2,6 +2,8 @@ require 'examples/configuration'
 
 require 'active_support/all'
 
-require 'examples/my_dice'
-require 'examples/my_dynamic_dice'
+# require 'examples/my_dice'
+# require 'examples/my_dynamic_dice'
+require 'marskal/module_configurator'
+require 'examples/dice_method_one'